Query: Adds an environment variable for disabling the hybrid search query plan optimization (#5120)
## Description
This change adds a configuration whose value will be read from the
environment variable
`AZURE_COSMOS_HYBRID_SEARCH_QUERYPLAN_OPTIMIZATION_DISABLED`. If this
evaluates to true, then the optimization for skipping the order by
rewrite for hybrid search is disabled.

The environment variable is read, and the value is propagated as an
internally visible query request option. This is plumbed down in the
query plan APIs where it is either used to populate the supported query
features header in case of a gateway or it is used as a flag passed into
ServiceInterop.

## Type of change
- [x] New feature (non-breaking change which adds functionality)