[Internal] Binary Encoding: Adds Client Option to Enable/Disable Request and Response Stream Conversation (#5123)
# Pull Request Template

## Description

Create a new internal cosmos client options: `EnableStreamPassThrough`
to enable or disable the request/ response stream conversation. When
enabled, the request and/or response stream would not be converted to
the desired target serialization type and would act just like a
pass-through. The default value for this parameter is `False`.

**Usage:**

```
      CosmosClientOptions clientOptions = new CosmosClientOptions()
      {
          EnableStreamPassThrough = true,
      };

      CosmosClient cosmosClient = new (
          endpoint,
          authKey,
          clientOptions);
```

## Type of change

Please delete options that are not relevant.

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)

## Closing issues

To automatically close an issue: closes #5122