[INTERNAL] AddressEnumerator: Fixes IndexOutOfRangeException Observed in AddressEnumerator  (#5126)
Due to a race condition between modifying and reading a shared variable
FailedEndpoints which is instantiated in the
DocumentServiceRequestContext class. multiple threads seems to be
accessing it at the same time - one adding to it in the
AddToFailedEndpoints() method in the DocumentServiceRequestContext and
the other in the AddressEnumerator where it is trying to read an item in
the GetEffectiveStatus() method which is called with in
MoveFailedReplicasToTheEnd() and it results in an
IndexOutOfRangeException . We need to move to a thread safe data
structure.

[] Bug fix (non-breaking change which fixes an issue)

## Closing issues

To automatically close an issue: closes
https://github.com/Azure/azure-cosmos-dotnet-v3/issues/5046

---------

Co-authored-by: Debdatta Kunda <87335885+kundadebdatta@users.noreply.github.com>