[Internal] Thin Client Integration: Fixes the ThinClientTransportSerializer to avoid creating memory stream. (#5137)
# Pull Request Template

## Description

Optimize the serialization logic by avoiding the creation of a new large
MemoryStream every time by switching to a pooled byte array using
ArrayPool<byte>.Shared. Instead of allocating a fresh buffer inside the
stream, now renting a reusable one, copying data into it, and wrapping
it with MemoryStream. This reduces memory usage.

## Type of change

Please delete options that are not relevant.

- [X] New feature (non-breaking change which adds functionality)

## Closing issues

To automatically close an issue: closes #5070

Co-authored-by: Debdatta Kunda <87335885+kundadebdatta@users.noreply.github.com>