[Internal] Thin Client Integration: Adds HTTP Client Sharing between Gateway and Thin Client (#5141)
# Pull Request Template

## Description

Adding HttpClient sharing logic between Gateway and Thinclient model.
HttpClient is capable of handling both Http1.1 and Http2.0 protocol. The
request will decide which protocol to use based on the connection mode.

Therefore, ideally it should be possible to share the same client with
both the (Gateway and Thin Client) store models.

## Type of change

Please delete options that are not relevant.

- [x] New feature (non-breaking change which adds functionality)

## Closing issues

To automatically close an issue: closes #4579