[Internal] PPAF: Adds Code to Enable Read Hedging By Default (#5165)
# Pull Request Template

## Description

This PR does the following two things:

- Enables Cross Regional Hedging by default with a threshold of either
`1` second or half of the request timeout value provided, whichever is
minimum, when PPAF is enabled. Note that the default availability
strategy can be opted out by setting the `AvailabilityStrategy` to
`DisabledAvailabilityStrategy` in the `CosmosClientOptions`. To
understand this better, please take a look at the example below:

CosmosClientOptions cosmosClientOptions = new CosmosClientOptions()
            {
                EnablePartitionLevelFailover = true,
                ConsistencyLevel = Cosmos.ConsistencyLevel.Strong,
AvailabilityStrategy = new DisabledAvailabilityStrategy(),
            };

- As of today, during the cosmos client initialization, with PPAF
enabled we ensure that either the `ApplicationRegion` or
`ApplicationPreferredRegions` is present through a validation. Since we
now have established the concept of Effective Preferred Region, this PR
removes the above mandatory validation on `ApplicationRegion` or
`ApplicationPreferredRegions`, as in the if there are no preferred
regions passed explicitly, the .NET SDK can populate the effective
preferred regions.

## Type of change

Please delete options that are not relevant.

- [x] New feature (non-breaking change which adds functionality)

## Closing issues

To automatically close an issue: closes #5047, #5164