Fix warnings on GetProperty/SetProperty
On Attachment and Document. Bubbling up warnings to the class
fixes these because the code is not reachable at runtime. It still warns
at the callsite without this fix due to:
https://github.com/dotnet/runtime/issues/116032#issuecomment-2913545589

For now the fix makes Attachment and Document wholly trim/aot-incompatible.
In the future we might want to annotate it in a way that allows some parts to
be used without warnings.