[Internal] Fix: Fixes aggressive connection timeout behavior test flakiness (#5245)
* Reduce the number of iterations of point creates for catching the
unhealthy channel properly.
 * Checked the `DeepHealthy` boolean to ascertain health of a connection
* Fine tuned request timeout values

# Pull Request Template

## Description

Please include a summary of the change and which issue is fixed. Include
samples if adding new API, and include relevant motivation and context.
List any dependencies that are required for this change.

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)
- [] New feature (non-breaking change which adds functionality)
- [] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [] This change requires a documentation update

## Closing issues

To automatically close an issue: closes #IssueNumber

Co-authored-by: Debdatta Kunda <87335885+kundadebdatta@users.noreply.github.com>
Co-authored-by: Nalu Tripician <27316859+NaluTripician@users.noreply.github.com>