[Internal] Gates: Fixes bug where CosmosOperaitonCancellationException is now thrown in CancellationTokenTests (#5076)
# Pull Request Template

## Description

In our pipelines, the most common flaky test is related to the
`CosmosOperationCanceledExceptionTests` class. Here, about 50% of the
time, an OperationCancelledException is thrown when the correct
exception is `CosmosOperationCanceledException`. This PR makes sure that
the correct exception is thrown in the `AbstractRetryHandler` class
where the exceptions originate from.

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)

## Closing issues

To automatically close an issue: closes #5208