[Internal] PPAF: Fixes Integration Tests to Intercept Gateway Response (#5261)
# Pull Request Template

## Description

Recently, we started observing in our CI pipeline gates that, the
gateway started returning the ppaf enablement flag as `False` (instead
of `null`) for the account used for end to end test. Below are the
significance of the values returned from the gateway:

| Flag    | GW Response | Description |
| -------- | ------- | ------- |
| `enablePerPartitionFailoverBehavior` | `True` | PPAF was enabled for
the account.
| `enablePerPartitionFailoverBehavior` | `False` | PPAF was manually
disabled for the account.
| `enablePerPartitionFailoverBehavior` | `null` | PPAF was never set for
the account.
| `enablePerPartitionFailoverBehavior` | flag not present | PPAF was
never set for the account.

This PR fixes the integration test by intercepting the
`enablePerPartitionFailoverBehavior` flag for the gateway response to
set it as `null`, so that the PPAF enablement environment variable can
be honored.

## Type of change

Please delete options that are not relevant.

- [x] Bug fix (non-breaking change which fixes an issue)

## Closing issues

To automatically close an issue: closes #IssueNumber