[Internal] Gate Reliability: Fixes flaky tests (#5259)
# Pull Request Template

## Description

This PR fixes several tests in the
`CosmosOperationCanceledExceptionTests` class. There is an issue with
these tests and how they compare the exception diagnostics/
`ToString()`, which can differ in some cases due to the time durations
are calculated and when other tests are run in parallel. This PR fixes
this issue by only checking for the start of the diagnostic string which
should be the same regardless of how it is calculated. The only thing we
need to validate with this test is whether the exception message
contains diagnostics, not the contents of the diagnostics itself.

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)

## Closing issues

To automatically close an issue: closes #IssueNumber

Co-authored-by: Kiran Kumar Kolli <kirankk@microsoft.com>