[Internal] ThinProxy Integration: Adds New `HttpTimeoutPolicy` and Fix Cross Regional Retry Logic (#5276)
# Pull Request Template

## Description

This PR adds the support for the below:

- **Adds new HttpTimeoutPolicy:** Primarily create a new http timeout
policy to retry the proxy requests faster than the regular Gateway
requests. We need an aggressive timeout policy to meet the SLA bar.

- **Adds/ Fixes Cross Regional Retry Logic:** Currently the cross
regional retry logic for thin client has a gap and it does not resolve
the thin proxy endpoints correctly. This PR is fixing that behavior.

## Type of change

Please delete options that are not relevant.

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)

## Closing issues

To automatically close an issue: closes #5275