[Internal] Thin Client Integration: Adds Thin Client Enablement Through Account Config and Reduce Request Headers for Proxy (#5303)
# Pull Request Template

## Description

This PR addresses the following:

- While creating the proxy request, the SDK will now pass the below
headers only:

  - `x-ms-user-Agent` : Used to identify the SDK user agent.
- `x-ms-activity-id` : Activity ID to track requests in the proxy layer.

- Update thin client enablement criteria:

Currently the TC enablement is the combination of the below two
operations:
    
    Enable TC = `Set Environment Variable` + `Select Gateway Mode`
     
If either of the above condition is not meeting, default is Gateway
mode. This PR is changing the enablement with the combination of `3`
criteria. The flow will look like the below:
     
First Phase Rollout: `Enable TC` = `Set Environment Variable` + `Select
Gateway Mode` + `GW Account Level Response indicating TC is enabled`
(containing TC endpoints)
     
If either of the above condition is not meeting, default is Gateway
mode.

## Type of change

Please delete options that are not relevant.

- [X] New feature (non-breaking change which adds functionality)

## Closing issues

To automatically close an issue: closes #5285
To automatically close an issue: closes #5282