[Internal] Thin Client Integration: Removes ThinClientStoreModel and uses GatewayStoreModel for better code modularity. (#5323)
# Pull Request Template

## Description

It is observed that most of the working principle on the
ThinClientStoreModel can be achieved by the GatewayStoreModel.
Therefore, removing ThinClientStoreModel and accommodating the logic to
navigate between thinclient mode and gateway mode in GatewayStoreModel.
Added tests to verify correct working of all three connection modes:
thinclient, gateway and direct.

## Type of change

Please delete options that are not relevant.

- [x] New feature (non-breaking change which adds functionality)


## Closing issues

To automatically close an issue: closes #5311