Query: Adds GetIndexMetrics LINQ extension method (#5313)
# Pull Request Template

## Description

This PR adds a new extension method `GetindexMetrics` that can be used
on a `Response<T>` object to get the index metrics information.
Normally, the index metrics is only accessible via
`FeedResponse.IndexMetrics`. This provide an additional way to obtain
this information when using LINQ async extension method like
CountAsync/SumAsync/AvgAsync, whose return type is a
`Response<TSource>`.

As part of this fix, this PR also address these async method not
populating `IndexUtilizationText` field on the `Response` object as
well.

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [] This change requires a documentation update

## Closing issues

To automatically close an issue: closes #IssueNumber

---------

Co-authored-by: leminh98 <leminh@microsoft.com>
Co-authored-by: Kiran Kumar Kolli <kirankk@microsoft.com>