[Internal] Thin Client Integration: Fixes 400/1041 Error for Query Operations in Thin Client Mode. (#5345)
# Pull Request Template

## Description
With the existing code, query requests are getting 400/1041 error
response which indicates that proxy is returning an error due to missing
headers. This is happening because it appears that the .NET SDK is not
sending the startepk and endEpk headers, as the resolved partition key
range is appearing to be null. So, these required headers are appearing
missing at the proxy layer and hence the operation fails.

Updated the check in GatewayStoreModel to ensure that it resolves
partitionKeyRange and added more tests to cover different consistencies
for query operations.

## Type of change

Please delete options that are not relevant.

- [X] Bug fix (non-breaking change which fixes an issue)

## Closing issues

To automatically close an issue: closes #5343

---------

Co-authored-by: Kiran Kumar Kolli <kirankk@microsoft.com>