[Internal] Query : Adds new test project for AOT (#5371)
# [Internal] Query : Adds new test project for AOT

## Description

This change adds a new test project for AOT purposes. Existing test
projects are not AOT compliant. They are also unable to reference the
AOT product dll, since much of the functionality has been removed (for
e.g. database and collection CRUD API, LINQ, ChangeFeed to name a few).
This generates over 1000 errors on the emulator test project alone and
it's not tenable at the moment to address those all in time.
The purpose of this new dll is to serve as a smoke test/validation. Over
time, it is expected to contain a subset of tests that are required to
have sufficient confidence in quality of AOT flavor of product code.

Since database/collection CRUD functionality is not available in the AOT
SDK, the test dll contains a ```RestHelper``` to facilitate common
management operations at database/collection level that can be utilized
for test setup/cleanup purposes.

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)
- [] New feature (non-breaking change which adds functionality)
- [] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [] This change requires a documentation update

## Closing issues

To automatically close an issue: closes #IssueNumber