TokenCredentialCache: Adds a fallback mechanism to AAD scope override. (#5364)
# Pull Request Template

## Description

It's a follow up for this PR
https://github.com/Azure/azure-cosmos-dotnet-v3/pull/5252, which
introduces the ability to override AAD scope value. As part of the
current PR, use overridden scope if its provided and there is no
fallback for this.
If no overridden scope is provided then use account scope and if this
fails with exception
[AADSTS500011](https://learn.microsoft.com/en-us/entra/identity-platform/reference-error-codes)
then fallback to this scope "https://cosmos.azure.com/.default". Each
scope continues to go through existing validations.
## Type of change

Please delete options that are not relevant.

- [X] New feature (non-breaking change which adds functionality)


## Closing issues

To automatically close an issue: closes #IssueNumber