[Internal]Nregion: Adds support for including nregioncommit synchronous for less than Strong Consistency (#5401)
## Description

Feature: N-region synchronous commit support for < Strong consistency in
SDK)

This PR brings in N region syncrhonous commit for accounts that have
this feature enabled. The feature primarly brings in Barrier requests
for Writes for less than Strong consistencies

The change contains two parts (one part is in this PR and other is done
in Msdata/released as Direct package version 3.41.0 which is consumed in
this PR)


Part 2 **THIS PR** : V3 Repo changes which include changes in
AccountProperties to read the support for Nregion Commit .

-Adds support for including nregioncommit account feature while creating
StoreClient within DocumentClient


Part 1 CHanges to ConsistencyWriter for adding Barrier requests and
StoreResult to support the new NRegionCommittedGLSN which is critical to
performing the barriers.

This is done in the Msdata ShareFiles and a new Direct package version
3.41.0 is released with this change

- Refactors ConsistencyWriter to support synchronous barrier commits in
NRegion accounts (< Strong) when EnableNRegionSynchronousCommit is
enabled.

- Ensures response is only returned once GlobalNRegionCommittedGLSN has
caught up with the write LSN, preserving linearizability

Change PR in msdata:
https://msdata.visualstudio.com/CosmosDB/_git/CosmosDB/pullrequest/1800184


## Direct package changes

Direct version upgrade to 3.41.0 brings in the following changes

- Part 1 from above: Adds write barrier logic to ConsistencyWriter using
StoreResult's N-region committed LSN to enforce N region synchronous
commit semantics for less-than-strong consistency writes.
- Introduced SDK support for new regions of SingaporeCentral and
SingaporeNorth
- Resolves failover handling in strong consistency writes by ensuring
stale barrier context is cleared when retrying across regions.
- Introduced new Length Aware Range Comparators in EPK Range class.

## Testing

Emulator test for NregionCommit Write Barrier

Used Http interceptor to "mock" nregioncommit feature enabled in
AccountProperties.
Used Transport interceptor to mock GLobalNRegionCommitLSN for testing
scenario.

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)
- [] New feature (non-breaking change which adds functionality)
- [] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [] This change requires a documentation update

## Closing issues

To automatically close an issue: closes #IssueNumber

---------

Co-authored-by: Debdatta Kunda <87335885+kundadebdatta@users.noreply.github.com>